Unify CardContent for senses from three places into single component.
