+++ src/goals/MergeDupGoal/MergeDupStep/DragDropComponents/SidebarDragSense.tsx
@@ -1,4 +1,4 @@
-import { CardContent, Typography, Grid, Chip, Card } from "@material-ui/core";
+import { Card } from "@material-ui/core";
 import { Draggable } from "react-beautiful-dnd";
 import { useSelector } from "react-redux";
 
@@ -7,6 +7,7 @@ import {
   MergeTreeSense,
   Sidebar,
 } from "goals/MergeDupGoal/MergeDupStep/MergeDupsTree";
+import SenseCardContent from "goals/MergeDupGoal/MergeDupStep/SenseCardContent";
 import { StoreState } from "types";
 import theme from "types/theme";
 
@@ -49,38 +50,13 @@ export default function SidebarDragSense(props: SidebarDragSenseProps) {
                 : "lightgrey",
             }}
           >
-            {senseCardContent(props.sense, showDefinitions)}
+            <SenseCardContent
+              senses={[props.sense]}
+              includeDefinitions={showDefinitions}
+            />
           </Card>
         </div>
       )}
     </Draggable>
-  );
-}
-
-function senseCardContent(
-  sense: MergeTreeSense,
-  showDefinitions: boolean
-): JSX.Element {
-  const sep = "; ";
-  return (
-    <CardContent>
-      <Typography variant={"h5"}>
-        {sense.glosses.map((g) => g.def).join(sep)}
-      </Typography>
-      <Grid container spacing={2}>
-        {sense.semanticDomains.map((dom) => (
-          <Grid item xs key={dom.name}>
-            <Chip label={`${dom.name} ${dom.id}`} />
-          </Grid>
-        ))}
-      </Grid>
-      {showDefinitions && (
-        <div style={{ background: "lightyellow", marginTop: theme.spacing(2) }}>
-          <Typography variant={"h6"}>
-            {sense.definitions.map((g) => g.text).join(sep)}
-          </Typography>
-        </div>
-      )}
-    </CardContent>
   );
 }
